+++ src/main/java/com/kestrel/weddingbookkeeper/api/wedding/facade/WeddingFacade.java
@@ -5,7 +5,7 @@ import com.kestrel.weddingbookkeeper.api.wedding.dto.response.WeddingIdResponse;
 
 public interface WeddingFacade {
 
-    void createWeddingInfo(Integer memberId, WeddingInfoRequest weddingInfoRequest);
+    void createWeddingInfo(Long memberId, WeddingInfoRequest weddingInfoRequest);
 
-    WeddingIdResponse getWeddingId(Integer memberId);
+    WeddingIdResponse getWeddingId(Long memberId);
 }
